Restructurado de frontend en js
