+++ Frontend/src/App.jsx
@@ -1,0 +1,8 @@
+import {RouterProvider} from 'react-router-dom'
+import router from './routes'
+
+export default function App(){
+    return (
+        <RouterProvider router={router} />
+    )
+}
