+++ Frontend/src/main.jsx
@@ -1,11 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import {RouterProvider} from 'react-router-dom'
-
-import router from './routes'
-
+import App from './App'
 ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
-      <RouterProvider router={router}/>
+        <App />
     </React.StrictMode>
 )
